Fix Validator

+++ backEnd/api/corsi/deleteCorso.php
@@ -6,7 +6,7 @@ class deleteCorso{
     public static $schema = [
         [
             'name' => "id",
-            "type" => "numeric",
+            "type" => "int",
             "required" => true
         ]
     ];
